All terminal entry don't have NULL anymore in their column child, but a 0x0 tibble. Closes #41
This is necessary because comments are top-level expressions that are at the same time terminals. Hence there is nothing that can be nested inside them. bind_rows(NULL, NULL) (in R/nested#69) does not give NULL but tibble 0x0, hence, the children of comments will be a 0x0 tibble, whereas for all other terminals, it was NULL. Adapting 0x0 tibble for all terinal children was therefore necessary to restore consistency.