add renesting for correct operator indention
This commits implements re-nesting of the parse table so that token between unary operators
such as +, -, /, * and SPECIAL (infix operators such as %>%) are all on the same level of nesting.
This is not the case after compute_parse_data_nested(). Since the indention of the tokens between
those operators is the same, changing the structure of the nested parse table seemed to be the
easiest way to correctly indent them.  Note that operator precedence (+,- vs. /,*) is ignored
here since it does not impact indention.