Fix indent indices computation with multiple braces
If braces to if-else statements and the result contained a brace which(pd$token %in% token_closing=[needs_indention]) yielded two numbers. Since the brace expression is not added in a new expression, but the braces are added around the current expression in a nest, the expression may contain pairs of braces.