fixing addin
Make sure the selection overwritten with the styled expression is the initial region. If the styled expression is shorter than the initial expression (and the reference is the length of the new expression), it may happen that some of the old code will still be below the newly inserted expression.