allow curly-curly arguments to be on its own line
E.g. call({{\nx }}\n) should be left as is. Because we can't look into the future 2 steps (token_after), we need to compute the token ad-hoc, check if it is {, and if yes, we know it is a curly-curly expression. Leaves the minor edge case of an expression starting with `{{` but not being a curly-curly expression (e.g. `call(\n{{1} + 2})`) treated inconsistently beause we assume here it is a curly curly expression (as we only check for `{{` in the beginning. Result, first `{` is moved to the new line.