make R.cache an import
This remove one step in the styler setup for caching. Now, the only thing an interactive user has to do is to allow R.cache to set up a permanent cache. I believe this is worth the three extra recursive dependencies because there is actually no reason why you are strictly against using caching, even on CI/CD it will speed up things.