mark also non-terminal tokens with stylerignore = TRUE if they lay in a stylerignore sequence.
Previously, all non-terminals were set to FALSE. This new implementation is required when creating new tokens (e.g. wrap_expr_in_curly()) and we don't want to make assumptions about the structure of the child to find out whether it's stylerignore or not. Hence, we can just check if the tokens to wrap are stylerignore or not, no matter whether or not they are terminals.